backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pydantic-settings
sqlalchemy
docling
qdrant-client
openai
google-generativeai
